+++ src/main/java/com/insane96mcp/shieldsplus/ShieldsPlus.java
@@ -3,6 +3,7 @@ package com.insane96mcp.shieldsplus;
 import com.insane96mcp.shieldsplus.setup.Client;
 import com.insane96mcp.shieldsplus.setup.Config;
 import com.insane96mcp.shieldsplus.setup.SPItems;
+import com.insane96mcp.shieldsplus.setup.SPRecipeSerializers;
 import com.mojang.logging.LogUtils;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModLoadingContext;
@@ -23,5 +24,6 @@ public class ShieldsPlus
         FMLJavaModLoadingContext.get().getModEventBus().addListener(Client::setup);
         final IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
         SPItems.ITEMS.register(modEventBus);
+        SPRecipeSerializers.RECIPE_SERIALIZERS.register(modEventBus);
     }
 }